Refactored logging, get_config, and get_connection methods to a new utils file.
